requirements.txt
streamlit==1.22.0
tensorflow==2.12.0
pillow
numpy
